Project_1/pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
